// apps/desktop/src/ipc/generated/EmbedderState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * The local embedding model's state.
 */
export type EmbedderState = { "type": "notInstalled" } | { "type": "downloading", received: number, total: number, } | { "type": "installed" } | { "type": "loading" } | { "type": "loaded" } | { "type": "failed", error: string, };